More/Default Options Configuration Changes to stop FoamFix from crashing on first game start
The B:initOptions in FoamFix stops More Default Options and Default Options from working when set to true. However, setting it to false causes a crash on a fresh install when using More Default Options to load the options.txt file.

The workaround is to use Default Options to load the options.txt file and keybinds, but More Default Options to load other files (e.g. Journey Map configs).